+++ cmd/internal/cli/keyserver.go
@@ -11,9 +11,7 @@
 package cli
 
 import (
-	"io"
 	"os"
-	"strings"
 
 	"github.com/apptainer/apptainer/docs"
 	"github.com/apptainer/apptainer/internal/app/apptainer"
@@ -172,25 +170,7 @@ func setKeyserver(_ *cobra.Command, _ []string) {
 var KeyserverLoginCmd = &cobra.Command{
 	Args: cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
-		loginArgs := new(apptainer.LoginArgs)
-
-		loginArgs.Name = args[0]
-
-		loginArgs.Username = loginUsername
-		loginArgs.Password = loginPassword
-		loginArgs.Tokenfile = loginTokenFile
-		loginArgs.Insecure = loginInsecure
-
-		if loginPasswordStdin {
-			p, err := io.ReadAll(os.Stdin)
-			if err != nil {
-				sylog.Fatalf("Failed to read password from stdin: %s", err)
-			}
-			loginArgs.Password = strings.TrimSuffix(string(p), "\n")
-			loginArgs.Password = strings.TrimSuffix(loginArgs.Password, "\r")
-		}
-
-		if err := apptainer.KeyserverLogin(remoteConfig, loginArgs); err != nil {
+		if err := apptainer.KeyserverLogin(remoteConfig, ObtainLoginArgs(args[0])); err != nil {
 			sylog.Fatalf("%s", err)
 		}
 	},