+++ internal/app/apptainer/remote_login.go
@@ -2,8 +2,8 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
+// Copyright (c) 2019-2023, Sylabs Inc. All rights reserved.
 // Copyright (c) 2020, Control Command Inc. All rights reserved.
-// Copyright (c) 2019-2021, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -76,9 +76,8 @@ func RemoteLogin(usrConfigFile string, args *LoginArgs) (err error) {
 		if args.Tokenfile != "" {
 			return fmt.Errorf("--tokenfile is only supported for login to a remote endpoint, not OCI (docker/oras) or keyservers")
 		}
-		if err := c.Login(args.Name, args.Username, args.Password, args.Insecure); err != nil {
-			return fmt.Errorf("while login to %s: %s", args.Name, err)
-		}
+		sylog.Warningf("'remote login' is deprecated for registries or keyservers and will be removed in a future release; running 'registry login'")
+		return RegistryLogin(usrConfigFile, args)
 	}
 
 	// truncating file before writing new contents and syncing to commit file
@@ -102,7 +101,7 @@ func RemoteLogin(usrConfigFile string, args *LoginArgs) (err error) {
 	return nil
 }
 
-// endPointLogin implements the flow to set a new token against a remote endpoing config.
+// endPointLogin implements the flow to set a new token against a remote endpoint config.
 // A token may be provided with a file, or through interactive prompts.
 func endPointLogin(ep *endpoint.Config, args *LoginArgs) error {
 	var (